Allowed the delta type to easily be changed.

+++ src/util/point.rs
@@ -11,9 +11,9 @@ pub struct Point<T> {
 
 impl Point<ComplexFixed<f32>> {
     // start_delta is the delta from reference of point (0, 0) in the image
-    pub fn new(image_x: usize, image_y: usize, image_width: usize, resolution: f64, top_left_delta: ComplexFixed<f32>) -> Self {
+    pub fn new(image_x: usize, image_y: usize, image_width: usize, resolution: f32, top_left_delta: ComplexFixed<f32>) -> Self {
         Point {
-            delta: ComplexFixed::new(image_x as f32 * resolution as f32 + top_left_delta.re, image_y as f64 * resolution + top_left_delta.im),
+            delta: ComplexFixed::new(image_x as f32 * resolution + top_left_delta.re, image_y as f32 * resolution + top_left_delta.im),
             index: image_y * image_width + image_x,
             iterations: 0,
             smooth: 0.0,
